+++ app/lib/widget/weight_picker.dart
@@ -1,9 +1,14 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:phosphor_flutter/phosphor_flutter.dart';
 import 'package:provider/provider.dart';
 import 'package:quantumphysique/quantumphysique.dart';
 
+import 'package:trale/core/constants.dart';
+import 'package:trale/core/l10n_extension.dart';
+import 'package:trale/core/text_size.dart';
 import 'package:trale/core/trale_notifier.dart';
 import 'package:trale/core/unit_precision.dart';
 import 'package:trale/core/units.dart';
@@ -13,3 +18,4 @@ import 'package:trale/core/units.dart';
 part 'weight_picker_parts/scroll_physics.dart';
 part 'weight_picker_parts/ruler_picker.dart';
 part 'weight_picker_parts/weight_slider.dart';
+part 'weight_picker_parts/weight_value_field.dart';